MODIFY: Correção na função de excluir estante, para verificar a dependencia

+++ model/controllers/controler_estante.py
@@ -167,18 +167,26 @@ class Estante:
     # Conexao com o banco de dados para excluir uma estante
     def remover_estante(cod_estante):
 
-        conexao = Conection.create_connection()
+        # Verifica se a estante possui uma dependencia 
+        if  Estante.verificar_dependencia_estante(cod_estante):
+            # Retorna se a remoção falhou por conta da dependência
+            return False # Não pode excluir
 
+        # Se não possuir uma dependencia, executa a exclusão da estante
+        conexao = Conection.create_connection()
         cursor = conexao.cursor()
 
         sql = "DELETE FROM estante WHERE cod_estante = %s;"
 
-        cursor.execute(sql, (cod_estante,))
+        valor = (cod_estante,)
+
+        cursor.execute(sql, valor)
+
         conexao.commit()
         
         cursor.close()
         conexao.close()
-        return True 
+        return True
     
     # Recupera as estantes registradas anteriormente
     def recuperar_estante(cpf):
